+++ src/main/java/org/ligerbots/robot2019/Main.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot;
+package org.ligerbots.robot2019;
 
 import edu.wpi.first.wpilibj.RobotBase;
