+++ src/main/java/org/ligerbots/robot2019/Robot.java
@@ -5,13 +5,13 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot;
+package org.ligerbots.robot2019;
 
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.Subsystems.DriveTrain;
-import frc.robot.Subsystems.Elevator;
+import org.ligerbots.robot2019.Subsystems.DriveTrain;
+import org.ligerbots.robot2019.Subsystems.Elevator;
 
 /**
  * The VM is configured to automatically run this class, and to call the
