+++ src/main/java/org/ligerbots/robot2019/RobotMap.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package org.ligerbots.robot2019;
 
 
 public class RobotMap {
