+++ src/main/java/org/ligerbots/robot2019/Subsystems/Elevator.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.Subsystems;
+package org.ligerbots.robot2019.Subsystems;
 
 import java.util.Arrays;
 
@@ -15,7 +15,7 @@ import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 
 import edu.wpi.first.wpilibj.PIDController;
 import edu.wpi.first.wpilibj.command.Subsystem;
-import frc.robot.RobotMap;
+import org.ligerbots.robot2019.RobotMap;
 
 /**
  * Add your docs here.
